additional sanity checks for model config
